+++ src/elements/db/OrderQuery.php
@@ -1261,7 +1261,7 @@ class OrderQuery extends ElementQuery
         }
 
         if (isset($this->email) && $this->email) {
-            $this->subQuery->andWhere(Db::parseParam('commerce_orders.email', $this->email));
+            $this->subQuery->andWhere(Db::parseParam('commerce_orders.email', $this->email, '=', true));
         }
 
         // Allow true ot false but not null